Better Certificate Checking
BOSH + gnatsd need to agree on names for the certificates that will be
used when dealing with nats over TLS.  The nats job configuration
hardcodes names that MUST be present as the subject of each of the
server / client / hm nats certs.

Now we check those assertions explicitly during `hooks check`, to make
sure we don't deploy unusable certificates.