New App Auto-Scaler Feature
This PR adds the ability for an operater to co-locate an App Autoscaler as a service broker to dynamically scale apps depending on policies provided.

The feature is `autoscaler`, but also brings along with it some general kit-related changes to make it easier for operators to deploy the App Autoscaler instance. Mostly, this relates to DB selection as AA requires a PostgreSQL DB. How AA selects a DB to use is outlined in MANUAL.md.

Documentation has been written, and this has been tested in vSphere with internal (HA/non-HA) PostgreSQL, external MySQL, and external PostgreSQL options chosen. All successfully deploy.

This release also relies on a non-yet-released boshrelease and as such relies on a self-hosted S3 blob with a dev tag.