Fix locketdb and silkdb writing to 'postgres' database (#40)
We had an incorrect spruce operation that confused `scheme` and
`schema`, which resulted in locketdb and diegodb data being written to a
database named `postgres` (rather than `locketdb` and `diegodb`,
respectively.)

This commit fixes this. However, it is not a backwards compatible
commit. If there is a current CF running that was deployed from CF
Genesis Kit versions 1.1.0 to 1.1.2, data migration is necessary.

There is a Genesis Migration Process document to accompany this fix:
http://www.genesisproject.io/docs/migrations/gmp-cf-0001