Release Note automation improvements
Added script to automate generation of upstream.yml that scans for all
releases, and populates the file.  Still a separate process from
building the release notes, as manual intervention may be required for
adding notes, fixing urls, etc...

release note generation now checks for tags with and without a prefix of
v, as the projects are inconsistant.