1.9.0 release candidate (#105)
* Releases updated to match cf-deployment 12.20.0

Also changes the health check path for uaa to
/var/vcap/jobs/uaa/bin/dns/healthy, as that was the only other change
between 12.5.0 and 12.20.0

* Fixed dns-service-discovery and haproxy certs location

* Improve test coverage of features

* Move smoke-tests to api instance-group

* BBS use local locket to prevent race conditions

* Increase timeout for deploying cell instances

* Moved smoketests to uaa from api for faster runs

* Added v1.9.0 release Notes

* Updated manual and network for HAProxy feature

HAProxy now uses the `haproxy_ips` feature, and can use it own separate
network, though it still defaults to cf_edge_network.  Updated manual to
reflect this.

* Refactor check hook after PR review

* Revert moving the haproxy ssl certs

Since the location is the same for the provide or self-signed certs that
are generated by Genesis, this falls out of the fully generated scope of
the other certs.