Make release table readable in terminal
The `genesis list-kits -u` prints out the release notes of upgrade kits
(or --details option does the same for all kits listed), but the release
table for core components and buildpacks is unreadable this way.

This makes the table fixed width so it maintains readability both in
html and on the terminal.