Allow for download and upload of stemcell files
`upload-stemcells` works fine when the bosh director could reach the
internet (except for the bugs in the last release).  However, if the
bosh director didn't have access to the internet due to network
configuration, intentional or otherwise, you couldn't get stemcells onto
it.

This commit adds a --dl flag to download the selected or specified
stemcells and then upload the local file.  It also added --dry-run if
you just want to show the url and sha1 for the files for use downloading
it in another manner.

In addition, to support mass-uploading to different directors without
mass-downloading, `download-stemcells` was added with the same selection
or specification interface as upload-stemcells, so you could download a
set of stemcells.  `upload-stemcells` was also updated to accept a
relative file location as a candidate to upload to allow you to upload
the downloaded stemcell files.