Transitioned RN gist to git repo
The storage of the Release Notes as a gist worked fine, but gists are
only editable by the user who created them (not available for orgs).
Moving to a genesis-community org repo to hold the pre-edited release
notes.

In shipit, we detect that the release notes are for the correct version
now, and error if they are not.  If it is a match, we purge the raw
commits and version trace comment before using it as the release body.

Also fixed some typos