v2.3.0 for nats migration and last bionic
[Migration Release]

The purpose of this release is to provide a step to migrate to v3.0.0.
This works by the following method:

* This version starts using the nats mbus certs instead of username and
  password, which becomes mandatory in bosh v274.0.0.  These certs were
  installed several releases ago, but were not in use.  They must be in
  use before upgrading beyond this version

* Once deployed, each deployment under the bosh director will need to
  get recreated(`genesis <env> deploy --recreate`) in order to propagate
  the nats changes.

* You can then upgrade to v3.0.0, and then recreate the deployments
  under the director once again to finalize the upgrade.

This is also the last verison to support bionic stemcells.  It also
supports jammy, using the feature `jammy` to turn it on.  Unfortunately,
there are no precompiled releases other than bosh-v273.1.0 for bionic,
so those will be built from source if using the default bionic
configuration.