Allow runtime config to be sourced
This change allows the runtime config to be sourced by other scripts
without executing the main logic. This is useful for defining routines
that can be used by other scripts.

The diff looks massive, but its really only a check for being sourced
and then some indentation changes, which confuses the diff tool to make
it look more massive.