Reverting using entombification temporarily
Entombification works, but previous version of the kit breaks it.  The
bug has been fixed, but because the upgrade test deployes the last good
version, it doesn't work with entombification turned on.

This will be reverted once the next version bosh kit is released, as
that will become the new last good version.