[ci] Dynamically build test jobs based on environment settings
This change introduces a new script `build-test-jobs` that will
dynamically build test jobs based on the environment settings. This
allows for more flexibility in the pipeline configuration and reduces
the need to deviate from the upstream pipeline template.

The script will read the `ci/envs/ci-<iaas>-*.yml` files and build test
jobs based on the `ci.job` and `kit.ci.type` settings. If the `kit.ci.type` is
`custom`, the script will look for a corresponding file in
`ci/custom-jobs` and copy it to the `pipeline/tests` directory. Otherwise,
it will create jobs in the pipeline/tests directory based on the `kit.ci.type`
of `deploy` or `upgrade`.

The script will also build a `_test_group.yml` file that will be used to
determine the order of the test jobs in the pipeline. This file will
contain a list of all the test jobs and the jobs that they depend on.

The `prepare` job in the pipeline will be updated to use the list of
test jobs to determine which jobs to wait for before running the
`prepare` job.  Any job that should not block the `prepare` job should
have `blocking: false` set in the `kit.ci` section of the environment file.

Environments that are not ready to be tested can be skipped by setting
`skip: true` in the `kit.ci` section of the environment file.

Tests that should not run in parallel with other tests can be grouped
together by setting the `serial_groups` setting in the `kit.ci` section.

WIP:  There is also intended support for ordering of the test jobs based
on the `after` setting in the `kit.ci` section of the environment file,
but this is not yet implemented.

In addition to this change, the `build-upstream-jobs` script has been
updated to support building jobs based on the `meta.upstream.manifest`
setting in the settings.yml file. This will create a job under the upstream
group that will fetch the latest upstream manifest for kits that utilize
upstream manifests.