[ci] Add a release cache to speed up sha1 checks
This change adds a release cache to the build-kit and ship-release jobs
to speed up sha1 checks. The cache is a JSON file that stores the
timestamp, etag, size, and sha1 of the release artifacts. If the cache
is present and the release artifact has not changed, the sha1 check is
is done against the cache instead of fetching the artifact again.