Refactor use_create_env processing
OCFP management environments are now gated by the presence of the
use_create_env parameter in the environment file. This is a more
explicit way to gate the creation of management environments than
the previous method of checking the environment name for "-mgmt" and
turning on the `proto` feature (which would not actually deploy as a
create-env under certain circumstances).