Improved OCFP OpenStack manifest fragments
This commit improves the OpenStack manifest fragments for the OCFP
deployment. It adds a inline method of setting values based on the scale
of the environment, and adds a disk size parameter to the `proto`
manifest fragment.

It supports changing the subnet/availability zone used to boot the BOSH
VMs from the environment parameters, rather than hardcoding it to
ocfp-0/z1.

It also sets the `boot_from_volume` and `root_disk.size` properties in
multiple places to ensure that the BOSH VMs are created with the
correct disk size.  THIS SHOULD BE RESOLVED TO A SINGLE LOCATION IN THE
FUTURE!