Update mgmt bosh for openstack for ocfp cpi values
This change updates the mgmt bosh for openstack to include ocfp cpi
values. This is a required change to support the ocfp cpi configuration.

This relies on recent changes to Genesis::Hooks::Blueprint to support
the `remove_files` methods, as it removes the base $cpi configs and
replaces them with the ocfp cpi configs, which take values from the ocfp
cpi config in vault instead of from the params key in the environment
file.