If terminate fails, entry Pry if it exists
If the Pry module is installed, the script will enter a Pry session if the
terminate function fails. This allows for debugging and inspection of the
environment at the point of failure.  Otherwise, the script will print an error
message and exit with a non-zero status.  This fixes a bug where the
entire script would fail if the Pry module didn't exist because `use` is
a pragma that gets run at compile time, not run time.