Update cloud-config hook to use dynamic cloud properties
The `cloud-config.pm` hook now dynamically determines cloud properties
for the `bosh-lb` VM extension based on the IaaS provider. This replaces
the static `cloud_properties` key with a call to the new method
`_cloud_properties_for_iaas`.

This change improves flexibility and reduces hardcoding in cloud
configuration generation.  Before it would put an invalid configuration
for cloud_properties that included an IaaS key with the actual
properties, now it will use the IaaS key only when needed.