Fix cloud-config.pm vm_extension definition
The previous commit erroneously tried to replace the cloud_properties
definition with a method call, which was done incorrectly, but more
importantly, did not understand what was being done by the builder
code.

The original problem was that the code specified `cloud_properties`
which is used verbatim, but if the key is `cloud_properties_for_iaas`
then the value is run through a method that returns a hashref
with the cloud properties for the specific IaaS by the builder.